add bookshelf remove __main__ fix assistant
