--- src/__main__.py
@@ -1,9 +1,0 @@
-# vim:fenc=utf-8
-#
-# Copyright © 2023 sakakibara <sakakibara@dyana>
-#
-# Distributed under terms of the MIT license.
-from cli import cli
-
-if __name__ == "__main__":
-    cli()
